+++ insteonrf/__init__.py
@@ -13,7 +13,7 @@ from .packet import (
     pkt_crc,
 )
 
-__version__ = "2.4.1"
+__version__ = "2.4.2"
 __all__ = [
     "Address",
     "Command",